Move Halfedge_and_face_graph_property_maps.h to BGL.
